refactor: type 관련 파일 구조 변경
