+++ packages/cli/src/types/enums.ts
@@ -1,0 +1,4 @@
+export enum CodingSite {
+  LeetCode = 'LeetCode',
+  Programmers = 'Programmers',
+}
